Remove a bunch of redundant stuff from news, move DynamicPage into its own file.
