+++ src/pages/common/DynamicPage.js
@@ -1,0 +1,173 @@
+import React from 'react'
+
+import BlackBoxArrow from '../../../img/black box arrow.png'
+
+const CompLabel = (props) => (
+  <div className='complabel-component'>
+    Component: <b>{props.label}</b>
+  </div>
+)
+
+const Cover = (props) => (
+  <div className='cover-component'>
+    <CompLabel label='Cover' />
+    <h1 className='cover-head'>
+      {props.head}
+    </h1>
+    <p className='cover-byline'>
+      {props.byline}
+    </p>
+  </div>
+)
+
+const Copy = (props) => (
+  <div className='copy-component'>
+    <CompLabel label='Copy' />
+    <div>{props.content}</div>
+  </div>
+)
+
+const Channels = (props) => (
+  <div className='channels-component'>
+    <CompLabel label='Channels' />
+    <div>{props.children}</div>
+  </div>
+)
+
+const Twitter = (props) => (
+  <div className='twitter-component'>
+    <CompLabel label='Twitter' />
+    <div>
+      Twitter:&nbsp;
+      <a href={`https://twitter.com/${props.handle}`} target='_blank' rel='noopener noreferrer'>
+        @{props.handle}
+      </a>
+    </div>
+  </div>
+)
+
+const Link = (props) => (
+  <div className='newsletter-component'>
+    <CompLabel label='Link' />
+    <div>
+      Google:&nbsp;
+      <a href={props.subscribeUrl} target='_blank' rel='noopener noreferrer'>
+        {props.subscribeUrl}
+      </a>
+    </div>
+  </div>
+)
+
+const Image = (props) => (
+  <div className='Image-component'>
+    <CompLabel label='Image' />
+    <div>
+      <img src={props.src} />
+    </div>
+  </div>
+)
+
+export const DateBox = (props) => (
+  <div className='Datebox-component' style={{ margin: '1rem' }}>
+    <div className='row'>
+      <div className='col-md-2 col-xs-12'>
+        <div style={{ float: 'left', width: '90%', height: '200px', marginLeft: '1rem', border: '1px solid rgba(0, 0, 0, .2)', textAlign: 'center', 'text-transform': 'uppercase' }}>
+          <span style={{ fontSize: '84px', color: '#4D4D4D', letterSpacing: '0px' }}>{props.day}</span>
+          <br />
+          <span style={{ fontSize: '52px', color: '#4D4D4D', letterSpacing: '0px' }}>{props.month}</span>
+        </div>
+      </div>
+      <div className='col-md-9 col-xs-12' style={{ overflow: 'hidden', marginLeft: '1rem', marginRight: '1rem' }}>
+        <span style={{ fontSize: '35px', color: '#000', letterSpacing: '0px' }}><b>{props.title}</b></span>
+        <p style={{ fontSize: '15px', margin: '0.5rem 0' }}>{props.date} {props.time}</p>
+        <p style={{ fontSize: '18px' }}>{props.description}</p><br />
+        <span style={{ fontSize: '18px', color: '#000', letterSpacing: '0px' }}>{props.signupText} <a href={props.link} style={{ color: '#523178' }}>{props.link}</a></span>
+      </div>
+    </div>
+  </div>
+)
+
+const NewsArticle = (props) => {
+  return (
+    <div style={{ textAlign: 'left', cursor: 'pointer' }}>
+      <div>
+        <img src={props.imageSource} style={{ position: 'relative', top: '20px', right: '65px', 'max-width': '70%', 'box-shadow': '0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19)', margin: '0px 0px 0px 10%' }} />
+      </div>
+      <div className='NewsArticle marginAuto-xs' style={{ textAlign: 'left', 'border-style': 'solid', padding: '1rem', borderColor: 'black', borderWidth: '1px', width: '70%', height: '370px' }} onClick={() => props.link && window.open(props.link, '_blank')}>
+        <br />
+        <div style={{ maxHeight: '285px', overflow: 'hidden' }}>
+          <b style={{ fontSize: '28px', letterSpacing: '0px', color: '#000' }}>{props.head}</b><br />
+          <p style={{ fontSize: '13px', textAlign: 'left', letterSpacing: '0px', bottom: '0', 'margin-top': '0.5rem' }}>by {props.author} <span style={{ fontSize: '13px', letterSpacing: '0px', float: 'right' }}> {props.date}</span></p>
+          <br />
+
+          <div style={{ fontSize: '18px', letterSpacing: '0px' }}> {props.byline.substring(0, 200) + '...'}</div>
+        </div>
+        <div style={{ textAlign: 'right' }}><img style={{ 'max-width': '10%', margin: '1rem', left: '50%' }} src={BlackBoxArrow} /></div>
+      </div>
+    </div>
+  )
+}
+
+const SpacerColumn = (props) => (
+  <div />
+)
+
+const stock = {
+  Channels,
+  Copy,
+  Cover,
+  Link,
+  Twitter,
+  Image,
+  NewsArticle,
+  SpacerColumn,
+  DateBox
+}
+
+class Section1 extends React.Component {
+  get root () {
+    return this.props.root
+  }
+
+  get rSection () {
+    return this.createElement(this.root)
+  }
+
+  createElement (def) {
+    return React.createElement(
+      stock[def.component],
+      def.props,
+      ((def.props || {}).children || [])
+        .map(c => this.createElement(c))
+    )
+  };
+
+  render () {
+    return (
+      <div className={this.root.props.sizeName}>
+        {this.rSection}
+      </div>
+    )
+  }
+}
+
+export default class DynamicPage extends React.Component {
+  get sections () {
+    return this.props.page.sections
+  }
+
+  get rSections () {
+    return this.sections.map(
+      s => <Section1 key={s.id} root={s} />
+    )
+  }
+
+  render () {
+    return (
+
+      <>
+        {this.rSections}
+      </>
+    )
+  }
+}
